Create placeholder action for testing recalibration package compilation (#819)
This is a placeholder action file. Manually triggered actions can't be run against other branches unless they already exist in the master branch. 

I'm going to use this so I can test [an action in a branch that's not yet ready for PR](https://github.com/metaspace2020/metaspace/blob/feat/recalibration/.github/workflows/msi_recal_publish.yml). 

I'm going to merge this without review, as this change shouldn't affect anything, and will be overwritten by a properly-reviewed PR when this feature is ready.